Fix locale selector on /unlocalized
